refactor(tests): adapt fixtures
